requirements.txt
psycopg2-binary==2.9.9; python_version < '3.13'
psycopg[binary]==3.2.12; python_version >= '3.13'
python-dotenv==1.0.0
supabase==2.3.4
openpyxl==3.1.2
selenium==4.16.0
requests==2.31.0
schedule==1.2.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.9.2
python-multipart==0.0.6
pywin32>=306; platform_system == "Windows"